examples/prod_fastapi_demo/pyproject.toml
[tool.poetry]
name = "gino-fastapi-demo"
version = "0.1.0"
description = "A demo of a production-ready project based on GINO and FastAPI."
authors = ["Aobo Shi <aobo.shi@decentfox.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.54.0"
gino-starlette = {path = "../.."} # gino = { version = "^1.0", extras = ["starlette"] }
importlib_metadata = { version = "^1.3.0", python = "<3.8" }
uvicorn = "^0.11.2"
gunicorn = "^20.0.4"
alembic = "^1.4.0"
psycopg2 = "^2.8.4"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"

[tool.poetry.plugins."gino_fastapi_demo.modules"]
"index" = "gino_fastapi_demo.views.index"
"users" = "gino_fastapi_demo.views.users"

[tool.black]
line-length = 80

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
